// src/pages/replica/battle/types.ts
enum Boon {
    AbilityIgnition = "abilityIgnition",
    AbsoluteDefense = "absoluteDefense",
    AttackIgnition = "attackIgnition",
    Barrier = "barrier",
    BarrierII = "barrierII",
    Berserk = "berserk",
    BerserkII = "berserkII",
    Blitzball = "blitzball",
    BloodlustX = "bloodlustX",
    Boost = "boost",
    BoostII = "boostII",
    Brave = "brave",
    BraveII = "braveII",
    BreakDefenseUp = "breakDefenseUp",
    BreakImmunity = "breakImmunity",
    BurstMode = "burstMode",
    ChargingAttack = "chargingAttack",
    CleavingAttack = "cleavingAttack",
    CriticalResistUp = "criticalResistUp",
    DamageAbsorption = "damageAbsorption",
    DamageReduction = "damageReduction",
    DarkBoost = "darkBoost",
    DarkWeapon = "darkWeapon",
    Darkforce = "darkforce",
    DivineShield = "divineShield",
    Drain = "drain",
    DriveIgnition = "driveIgnition",
    EarthBoost = "earthBoost",
    EarthOfChaos = "earthOfChaos",
    EarthPact = "earthPact",
    EarthWeapon = "earthWeapon",
    Earthforce = "earthforce",
    EnhanceElementalAttacks = "enhanceElementalAttacks",
    EnhanceElementalAttacksII = "enhanceElementalAttacksII",
    EternalLight = "eternalLight",
    Faith = "faith",
    FaithII = "faithII",
    FeatherCharge = "featherCharge",
    FireBoost = "fireBoost",
    FirePact = "firePact",
    FireWeapon = "fireWeapon",
    Flameforce = "flameforce",
    FlamesOfChaos = "flamesOfChaos",
    FullRemission = "fullRemission",
    GiantsDrink = "giantsDrink",
    Haste = "haste",
    Iceforce = "iceforce",
    ImmuneToAilments = "immuneToAilments",
    ImmuneToBio = "immuneToBio",
    ImmuneToBreakDefenseDown = "immuneToBreakDefenseDown",
    ImmuneToCriticalResistDown = "immuneToCriticalResistDown",
    ImmuneToCurse = "immuneToCurse",
    ImmuneToDebarrier = "immuneToDebarrier",
    ImmuneToDebilitate = "immuneToDebilitate",
    ImmuneToDebrave = "immuneToDebrave",
    ImmuneToSleep = "immuneToSleep",
    ImmuneToSlow = "immuneToSlow",
    ImmuneToStun = "immuneToStun",
    ImmuneToTaunt = "immuneToTaunt",
    ImmuneToUnguard = "immuneToUnguard",
    ImmuneToWeaknessDown = "immuneToWeaknessDown",
    ImmuneToFatigue = "immuneToFatigue",
    ImmuneToHex = "immuneToHex",
    KnightsOfTheRound = "knightsOfTheRound",
    LightBoost = "lightBoost",
    LightPact = "lightPact",
    LightWeapon = "lightWeapon",
    Lightforce = "lightforce",
    LucidCast = "lucidCast",
    LucidCastII = "lucidCastII",
    LucidFist = "lucidFist",
    LucidFistII = "lucidFistII",
    LucidHunt = "lucidHunt",
    LucidHuntII = "lucidHuntII",
    LucidWar = "lucidWar",
    LucidWarII = "lucidWarII",
    MartialArts = "martialArts",
    MartialFlow = "martialFlow",
    Melee = "melee",
    Omnidrive = "omnidrive",
    PearlOfDestruction = "pearlOfDestruction",
    PerfectDefense = "perfectDefense",
    Quicken = "quicken",
    Ranged = "ranged",
    Regen = "regen",
    ResistDark = "resistDark",
    ResistEarth = "resistEarth",
    ResistFire = "resistFire",
    ResistLight = "resistLight",
    ResistWater = "resistWater",
    ResistWind = "resistWind",
    RindBarrier = "rindBarrier",
    SeedCounter = "seedCounter",
    Snipe = "snipe",
    SnipeII = "snipeII",
    TheUltimaWeaponGuardianBuff = "theUltimaWeaponGuardianBuff",
    TranceAll = "tranceAll",
    TranceAllII = "tranceAllII",
    TriBarrierFire = "triBarrierFire",
    TriBarrierWater = "triBarrierWater",
    TriBarrierWind = "triBarrierWind",
    UltimateBooster = "ultimateBooster",
    UltimateCharge = "ultimateCharge",
    UltimateGauge = "ultimateGauge",
    UmbralVeil = "umbralVeil",
    Veil = "veil",
    Wall = "wall",
    WallII = "wallII",
    WardOfChaos = "wardOfChaos",
    WaterBoost = "waterBoost",
    WaterPact = "waterPact",
    WaterWeapon = "waterWeapon",
    WavesOfChaos = "wavesOfChaos",
    WeaknessWeapon = "weaknessWeapon",
    WindBoost = "windBoost",
    WindPact = "windPact",
    WindWeapon = "windWeapon",
    Windforce = "windforce",
    WindsOfChaos = "windOfChaos",
}

enum Ailment {
    AbilityBlock = "abilityBlock",
    Bio = "bio",
    BioII = "bioII",
    BleedingOut = "bleedingOut",
    BreakDefenseDown = "breakDefenseDown",
    BreakDefenseDownII = "breakDefenseDownII",
    CallOfTheVoid = "callOfTheVoid",
    CloudedCast = "cloudedCast",
    CloudedFist = "cloudedFist",
    CloudedHunt = "cloudedHunt",
    CloudedWar = "cloudedWar",
    CriticalResistDown = "criticalResistDown",
    CriticalResistDownII = "criticalResistDownII",
    Curse = "curse",
    CurseII = "curseII",
    DarkLockout = "darkLockout",
    DeathsCurse = "deathsCurse",
    Debarrier = "debarrier",
    DebarrierII = "debarrierII",
    Debilitate = "debilitate",
    Debrave = "debrave",
    DebraveII = "debraveII",
    EarthLockout = "earthLockout",
    EdictOfEarth = "edictOfEarth",
    EdictOfFire = "edictOfFire",
    EdictOfLight = "edictOfLight",
    Fatigue = "fatigue",
    FireLockout = "fireLockout",
    Foam = "foam",
    Genesis = "genesis",
    Hex = "hex",
    HPSap = "hpSap",
    LaserDamage = "laserDamage",
    LightLockout = "lightLockout",
    MelonSeed = "melonSeed",
    Nightmare = "nightmare",
    NullWall = "nullWall",
    NullifyRecovery = "nullifyRecovery",
    Oil = "oil",
    Rampage = "rampage",
    SelfDisintegration = "selfDisintegration",
    Sleep = "sleep",
    Slow = "slow",
    Slump = "slump",
    Stun = "stun",
    Taunt = "taunt",
    UltimateDrain = "ultimateDrain",
    UltimateGaugeLocked = "ultimateGaugeLocked",
    Unguard = "unguard",
    Vulnerability = "vulnerability",
    WaterLockout = "waterLockout",
    Weaken = "weaken",
    WeakenII = "weakenII",
    WindLockout = "windLockout",
}
// ...
interface Effect {
    name: Boon | Ailment;
    duration: number;
    target: Target;
    timing: "before" | "after";
    type?: "square" | "hexagon";
}